requirements.txt
streamlit
streamlit-js-eval
psycopg2-binary
supabase